move ConfigParams to new file
