+++ slk/create_config_files.py
@@ -18,16 +18,13 @@ The configs_dir (where the config files will reside) can be set through the comm
 or the environment variable RIPPLED_SIDECHAIN_CFG_DIR
 """
 
-import argparse
 import json
-import os
 import sys
 import traceback
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Dict, List, Optional, Tuple, Union
 
-from dotenv import load_dotenv
 from xrpl import CryptoAlgorithm
 from xrpl.core.addresscodec import decode_seed, encode_account_public_key
 from xrpl.core.addresscodec.codec import _FAMILY_SEED_PREFIX, SEED_LENGTH, _encode
@@ -37,9 +34,8 @@ from xrpl.wallet import Wallet
 from slk.chain import Chain, single_node_chain
 from slk.common import eprint, same_amount_new_value
 from slk.config_file import ConfigFile
+from slk.config_params import ConfigParams
 from slk.config_strs import get_cfg_str, get_ips_stanza
-
-load_dotenv()
 
 MAINNET_VALIDATORS = """
 [validator_list_sites]
@@ -381,68 +377,6 @@ def generate_multinode_net(
             full_history=True,
             with_hooks=False,
         )
-
-
-def parse_args():
-    parser = argparse.ArgumentParser(
-        description=("Create config files for testing sidechains")
-    )
-
-    parser.add_argument(
-        "--exe",
-        "-e",
-        help=("path to rippled executable"),
-    )
-
-    parser.add_argument(
-        "--usd",
-        "-u",
-        action="store_true",
-        help=("include a USD/root IOU asset for cross chain transfers"),
-    )
-
-    parser.add_argument(
-        "--cfgs_dir",
-        "-c",
-        help=(
-            "path to configuration file dir (where the output config files will be "
-            "located)"
-        ),
-    )
-
-    return parser.parse_known_args()[0]
-
-
-class ConfigParams:
-    def __init__(self):
-        args = parse_args()
-
-        if "RIPPLED_MAINCHAIN_EXE" in os.environ:
-            self.exe = os.environ["RIPPLED_MAINCHAIN_EXE"]
-        if args.exe:
-            self.exe = args.exe
-        # if `self.mainchain_exe` doesn't exist (done this way for typing purposes)
-        if not hasattr(self, "exe"):
-            raise Exception(
-                "Missing exe location. Either set the env variable "
-                "RIPPLED_MAINCHAIN_EXE or use the --exe_mainchain command line switch"
-            )
-
-        self.configs_dir = None
-        if "RIPPLED_SIDECHAIN_CFG_DIR" in os.environ:
-            self.configs_dir = os.environ["RIPPLED_SIDECHAIN_CFG_DIR"]
-        if args.cfgs_dir:
-            self.configs_dir = args.cfgs_dir
-        # if `self.configs_dir` doesn't exist (done this way for typing purposes)
-        if not hasattr(self, "configs_dir"):
-            raise Exception(
-                "Missing configs directory location. Either set the env variable "
-                "RIPPLED_SIDECHAIN_CFG_DIR or use the --cfgs_dir command line switch"
-            )
-
-        self.usd = False
-        if args.usd:
-            self.usd = args.usd
 
 
 def main(params: ConfigParams, xchain_assets: Optional[Dict[str, XChainAsset]] = None):
